test(config): close OpenShell bridge review gaps
