+++ src/lib/private-networks.ts
@@ -17,6 +17,8 @@ import { ROOT } from "./runner";
 
 const NETWORKS_FILE = path.join(ROOT, "nemoclaw-blueprint", "private-networks.yaml");
 
+export const OPENSHELL_SANDBOX_HOST_BRIDGE = "host.openshell.internal";
+
 export interface NetworkEntry {
   address: string;
   prefix: number;
@@ -189,6 +191,25 @@ export function isPrivateHostname(hostname: string): boolean {
     if (normalised === reserved || normalised.endsWith(`.${reserved}`)) return true;
   }
   return isPrivateIp(normalised);
+}
+
+export function isAllowedOpenShellSandboxBridgeUrl(url: URL): boolean {
+  const hostname =
+    url.hostname.startsWith("[") && url.hostname.endsWith("]")
+      ? url.hostname.slice(1, -1)
+      : url.hostname;
+  const port = Number(url.port);
+  return (
+    hostname.replace(/\.$/, "").toLowerCase() === OPENSHELL_SANDBOX_HOST_BRIDGE &&
+    url.protocol === "http:" &&
+    Number.isInteger(port) &&
+    port >= 1024 &&
+    port <= 65535 &&
+    !url.username &&
+    !url.password &&
+    !url.search &&
+    !url.hash
+  );
 }
 
 /**
